WIP - convex, missing polygon
