+++ src/adapters/curve-dex/polygon/index.ts
@@ -1,11 +1,10 @@
+import { getPoolsBalances } from '@adapters/curve-dex/common/balance'
+import { getPoolsContracts } from '@adapters/curve-dex/common/pool'
+import { getRegistries } from '@adapters/curve-dex/common/registries'
+import { getGaugesContracts } from '@adapters/curve-dex/commonz/gauges'
 import type { BaseContext, Contract, GetBalancesHandler } from '@lib/adapter'
 import { resolveBalances } from '@lib/balance'
 import type { Token } from '@lib/token'
-
-import { getGaugesBalances, getPoolsBalances } from '../common/balance'
-import { getGaugesContracts } from '../common/gauges'
-import { getPoolsContracts } from '../common/pools'
-import { getRegistries } from '../common/registries'
 
 const CRV: Token = {
   chain: 'polygon',
@@ -16,26 +15,31 @@ const CRV: Token = {
 
 const xChainGaugesFactory: Contract = {
   chain: 'polygon',
-  address: '0xabC000d88f23Bb45525E447528DBF656A9D55bf5',
+  address: '0xabc000d88f23bb45525e447528dbf656a9d55bf5',
 }
 
 export const getContracts = async (ctx: BaseContext) => {
-  const registries = await getRegistries(ctx, ['stableSwap', 'stableFactory', 'cryptoSwap', 'cryptoFactory'])
-  const pools = await getPoolsContracts(ctx, registries)
+  const registries = await getRegistries(ctx, ['stableSwap', 'stableFactory', 'cryptoSwap'])
+  const pools = await getPoolsContracts(ctx, [
+    registries.stableSwap!,
+    registries.stableFactory!,
+    registries.cryptoSwap!,
+  ])
+
   const gauges = await getGaugesContracts(ctx, pools, xChainGaugesFactory, CRV)
 
   return {
     contracts: {
-      gauges,
       pools,
+      gauges,
     },
   }
 }
 
 export const getBalances: GetBalancesHandler<typeof getContracts> = async (ctx, contracts) => {
   const balances = await resolveBalances<typeof getContracts>(ctx, contracts, {
-    pools: (ctx, pools) => getPoolsBalances(ctx, pools, undefined, true),
-    gauges: (...args) => getGaugesBalances(...args),
+    pools: getPoolsBalances,
+    // gauges: getGaugesBalances,
   })
 
   return {
